% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mutation_frequency_bin_region.R
\name{calc_mutation_frequency_bin_region}
\alias{calc_mutation_frequency_bin_region}
\title{Calculate Mutation Frequency By Sliding Window.}
\usage{
calc_mutation_frequency_bin_region(
  region,
  these_samples_metadata = NULL,
  these_sample_ids = NULL,
  this_seq_type = "genome",
  maf_data = NULL,
  projection = "grch37",
  slide_by = 100,
  window_size = 1000,
  return_format = "long",
  min_count_per_bin = 0,
  return_count = TRUE,
  drop_unmutated = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{region}{A string describing a genomic region in the "chrom:start-end" format.
The region must be specified in this format OR as separate chromosome, start_pos, end_pos arguments.}

\item{these_samples_metadata}{Optional data frame containing a sample_id column.
If not providing a maf file, seq_type is also a required column.}

\item{these_sample_ids}{Optional vector of sample IDs. Output will be subset
to IDs present in this vector.}

\item{this_seq_type}{Optional vector of seq_types to include in heatmap.
Default is "genome". Uses default seq_type priority for samples
with >1 seq_type.}

\item{maf_data}{Optional maf data frame. Will be subset to rows where
Tumor_Sample_Barcode matches provided sample IDs or metadata table.
If not provided, maf data will be obtained with get_ssm_by_regions().}

\item{projection}{Specify which genome build to use. Required. Default grch37.}

\item{slide_by}{Slide size for sliding window. Default 100.}

\item{window_size}{Size of sliding window. Default 1000.}

\item{return_format}{Return format of mutations. Accepted inputs are "long"
and "wide". Long returns a data frame of one sample ID/window per row.
Wide returns a matrix with one sample ID per row and one window per column.
Using the "wide" format will retain all samples and windows regardless of
the drop_unmutated or min_count_per_bin parameters.}

\item{min_count_per_bin}{Minimum counts per bin, default is 0. Setting this
greater than 0 will drop unmutated windows only when return_format is long.}

\item{return_count}{Boolean statement to return mutation count per window (TRUE)
or binary mutated/unmutated status (FALSE). Default is TRUE.}

\item{drop_unmutated}{Boolean for whether to drop windows with 0 mutations.
Only effective with "long" return format.}

\item{...}{Any additional parameters.}
}
\value{
Either a matrix or a long tidy table of counts per window.
}
\description{
Count the number of mutations in a sliding window across a
region for all samples.
}
\details{
This function is called to return the mutation frequency for a given
region, either from a provided input maf data frame or from the GAMBL maf data.
Regions are specified with the \code{region} parameter. Alternatively, the region of
interest can also be specified by calling the function with \code{chromosome},
\code{start_pos}, and \code{end_pos} parameters. This function operates on a single region.
To return a matrix of sliding window counts over multiple regions,
see \code{calc_mutation_frequency_bin_regions}.
}
\examples{
myc_region = "8:128747680-128753674" 
myc_mut_freq = calc_mutation_frequency_bin_region(region = myc_region,
                                                  slide_by = 10,
                                                  window_size = 10000)
dplyr::arrange(myc_mut_freq,desc(mutation_count))

}
