% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMBLR.data-package.R
\docType{package}
\name{GAMBLR.open-package}
\alias{GAMBLR.open}
\alias{GAMBLR.open-package}
\title{GAMBLR.open: GAMBLR.open}
\description{
Database, mining, manipulation and convenience functions for working with genomic data of mature B-cell lymphomas.
}
\author{
\strong{Maintainer}: Ryan Morin \email{rdmorin@sfu.ca} (\href{https://orcid.org/0000-0003-2932-7800}{ORCID})

Authors:
\itemize{
  \item Kostia Dreval (\href{https://orcid.org/0000-0002-6214-2843}{ORCID})
}

}
\keyword{internal}
