% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_coding_ssm.R
\name{get_all_coding_ssm}
\alias{get_all_coding_ssm}
\title{Get all Coding SSMs}
\usage{
get_all_coding_ssm(
  these_samples_metadata = NULL,
  include_silent = FALSE,
  projection = "grch37"
)
}
\arguments{
\item{these_samples_metadata}{Supply a metadata table containing the sample/seq_type combinations you want.}

\item{include_silent}{If set to TRUE, silent/synonymous mutations in the coding regions will also be returned.}

\item{projection}{The desired genome build
'grch37' or 'hg38' are allowed. Default is grch37}
}
\value{
A data frame containing all the MAF data columns (one row per mutation).
}
\description{
Retrieve all coding SSMs from GAMBL in MAF-like format, regardless of seq_type.
}
\details{
Effectively retrieve coding SSM calls from one or all DNA seq_type. For additional optional arguments, see \link[GAMBLR.results:get_coding_ssm]{GAMBLR.results::get_coding_ssm}
}
\examples{
suppressPackageStartupMessages(library(GAMBLR.open))
my_meta = get_gambl_metadata(seq_type_filter = c("genome","capture"))
my_meta = check_and_clean_metadata(my_meta,duplicate_action="keep_first")
maf_all_seqtype = get_all_coding_ssm(my_meta)

table(maf_all_seqtype$maf_seq_type)

# most common mutations by gene and Variant_Classification
dplyr::group_by(maf_all_seqtype,
                Hugo_Symbol,
                Variant_Classification) \%>\% 
  dplyr::count() \%>\% 
  dplyr::arrange(desc(n))
}
