% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssm_by_sample.R
\name{get_ssm_by_sample}
\alias{get_ssm_by_sample}
\title{Get SSM By Sample.}
\usage{
get_ssm_by_sample(
  this_sample_id,
  these_samples_metadata = NULL,
  this_seq_type = "genome",
  projection = "grch37",
  ...
)
}
\arguments{
\item{this_sample_id}{A single sample ID to retrieve mutations for.}

\item{these_samples_metadata}{Optional, a metadata table (with sample_id
column) to auto-subset the data to samples in that table before returning.}

\item{this_seq_type}{Default is genome.}

\item{projection}{The projection genome build. Supports hg38 and grch37.}

\item{...}{Any additional parameters passed to \link{get_ssm_by_samples}.}
}
\value{
data frame in MAF format.
}
\description{
Get the SSMs (i.e. load MAF) for a single sample.
}
\details{
Alias for \link{get_ssm_by_samples} accepting a single sample ID via
\code{this_sample_id}, provided for API compatibility with GAMBLR.results.
}
\examples{

maf = get_ssm_by_sample(this_sample_id = "DOHH-2")

}
