% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_targetscan.R
\name{target_targetscan}
\alias{target_targetscan}
\title{Integrate miRNA (MicroRNA) data.}
\usage{
target_targetscan(maf, mirna_target)
}
\arguments{
\item{maf}{Required. MAF data frame (required columns: Chromosome, Start_Position, End_Position).}

\item{mirna_target}{Data frame contain miRNA data (required columns: Chromosome, Start_Position, End_Position, miRNA)}

\item{projection}{The genome build projection for the variants you are working with (default is grch37)}
}
\value{
data frame in MAF format with mirna (and sites := length of the seed region) columns.
}
\description{
Add miRNA data to MAF file.
}
\details{
For each variant, this function checks its genomic position against known miRNA target sites.
When an overlap is found, it appends the matched miRNA name(s) and target site details to the variant.
}
